Add max_target_perl to control the maximum minimum perl that's OK
Fixes GH #1, reported by xdg++

+++ t/everything.t
@@ -1,29 +1,58 @@
 use strict;
 use warnings;
-use Test::More 0.96 tests => 1;
+use Test::More 0.96 tests => 2;
 use autodie;
 use Test::DZil;
 use Moose::Autobox;
 
-my $tzil = Builder->from_config(
-    { dist_root => 'corpus/DZ1' },
-    {
-        add_files => {
-            'source/dist.ini' => simple_ini(
-                ('GatherDir', 'Test::MinimumVersion')
-            ),
+subtest 'explicit version' => sub {
+    plan tests => 2;
+
+    my $tzil = Builder->from_config(
+        { dist_root => 'corpus/DZ1' },
+        {
+            add_files => {
+                'source/dist.ini' => simple_ini(
+                    ('GatherDir', ['Test::MinimumVersion' => { max_target_perl => '5.10.1' }])
+                ),
+            },
         },
-    },
-);
-$tzil->build;
+    );
+    $tzil->build;
+    END { # Remove (empty) dir created by building the dists
+        require File::Path;
+        File::Path::rmtree('tmp') if -d 'tmp';
+    }
 
-my @xtests = map $_->name =~ m{^xt/} ? $_->name : (), $tzil->files->flatten;
-ok(
-    (grep { $_ eq 'xt/release/minimum-version.t' } @xtests),
-    'minimum-version.t exists'
-) or diag explain \@xtests;
+    my ($test) = map { $_->name eq 'xt/release/minimum-version.t' ? $_ : () } $tzil->files->flatten;
+    ok $test, 'minimum-version.t exists'
+        or diag explain [ map { $_->name } $tzil->files->flatten ];
 
-END { # Remove (empty) dir created by building the dists
-    require File::Path;
-    File::Path::rmtree('tmp');
-}
+    like $test->content => qr{\Q5.10.1\E}, 'max_target_perl used in test';
+};
+
+subtest 'version from metayml' => sub {
+    plan tests => 2;
+
+    my $tzil = Builder->from_config(
+        { dist_root => 'corpus/DZ1' },
+        {
+            add_files => {
+                'source/dist.ini' => simple_ini(
+                    ('GatherDir', 'Test::MinimumVersion')
+                ),
+            },
+        },
+    );
+    $tzil->build;
+    END { # Remove (empty) dir created by building the dists
+        require File::Path;
+        File::Path::rmtree('tmp') if -d 'tmp';
+    }
+
+    my ($test) = map { $_->name eq 'xt/release/minimum-version.t' ? $_ : () } $tzil->files->flatten;
+    ok $test, 'minimum-version.t exists'
+        or diag explain [ map { $_->name } $tzil->files->flatten ];
+
+    like $test->content => qr{metayml}, 'metayml used in test';
+};
